### 2020-10-15 #65 @whedon
[PRE REVIEW]: ExaPF.jl: A Power Flow Solver for GPUs
**Submitting author:** @michel2323 (<a href="http://orcid.org/0000-0002-4164-027X">Michel Schanen</a>)
**Repository:** <a href="https://github.com/exanauts/JuliaCon2020" target ="_blank">https://github.com/exanauts/JuliaCon2020</a>
**Version:**
