### 2021-01-28 #65 @vchuravy
@michel2323 It's complaining that https://github.com/exanauts/JuliaCon2020 is no longer available.

Can you take a look at https://juliacon.github.io/proceedings-guide/author/ and https://github.com/JuliaCon/JuliaConSubmission.jl
You probably need to re-submit so that we get Whedon happy.

We recommend that you have a paper branch/folder in ExaPF.jl instead of having a separate repository.
