### 2021-01-28 #65 @michel2323
@vchuravy First, the repository was not public. So it was just sitting there. Sorry about that. Second, it's the first time I read these guidelines. I was following the JOSS submission guidelines. Thus we created a pandoc markdown document.

What I will try to do now is to get a tex from our markdown file, put it in the ExaPF.jl repository as described by the guide and clone that in the repository address linked here to avoid dealing with resubmission and whedon.
