### 2021-01-28 #65 @vchuravy
Resubmission should be fairly painless, but whatever works best for you.

>  Second, it's the first time I read these guidelines

Yeah sorry we messed up and accidentally hid the guide for the 2020 conference.
