+++ paper/ref.bib
@@ -1,0 +1,110 @@
+@article{bicgstabVorst,
+author = {van der Vorst, H. A.},
+title = {Bi-CGSTAB: A Fast and Smoothly Converging Variant of Bi-CG for the Solution of Nonsymmetric Linear Systems},
+journal = {SIAM Journal on Scientific and Statistical Computing},
+volume = {13},
+number = {2},
+pages = {631-644},
+year = {1992},
+doi = {10.1137/0913035},
+}
+@article{sleijpen1993bicgstab,
+  title={BiCGstab(L) for linear equations involving unsymmetric matrices with complex spectrum},
+  author={Sleijpen, Gerard LG and Fokkema, Diederik R},
+  journal={Electronic Transactions on Numerical Analysis.},
+  volume={1},
+  pages={11--32},
+  year={1993},
+  publisher={Kent State University}
+}
+@article{schwarz,
+  author={S. {Abhyankar} and E. M. {Constantinescu} and B. F. {Smith} and A. J. {Flueck} and D. A. {Maldonado}},
+  journal={IEEE Transactions on Smart Grid},
+  title={Parallel Dynamics Simulation Using a Krylov-Schwarz Linear Solution Scheme},
+  year={2017},
+  volume={8},
+  number={3},
+  pages={1378-1386},}
+
+@article{RevelsLubinPapamarkou2016,
+    title = {Forward-Mode Automatic Differentiation in {J}ulia},
+   author = {{Revels}, J. and {Lubin}, M. and {Papamarkou}, T.},
+  journal = {arXiv:1607.07892 [cs.MS]},
+     year = {2016},
+     url = {https://arxiv.org/abs/1607.07892}
+}
+@article{besard2018juliagpu,
+  author        = {Besard, Tim and Foket, Christophe and De Sutter, Bjorn},
+  title         = {Effective Extensible Programming: Unleashing {Julia} on {GPUs}},
+  journal       = {IEEE Transactions on Parallel and Distributed Systems},
+  year          = {2018},
+  doi           = {10.1109/TPDS.2018.2872064},
+  ISSN          = {1045-9219},
+  archivePrefix = {arXiv},
+  eprint        = {1712.03112},
+  primaryClass  = {cs.PL},
+}
+
+% specific paper on array programming for heterogeneous systems (incl. CuArrays.jl)
+@article{besard2019prototyping,
+  title         = {Rapid software prototyping for heterogeneous and distributed platforms},
+  author        = {Besard, Tim and Churavy, Valentin and Edelman, Alan and De Sutter, Bjorn},
+  journal       = {Advances in Engineering Software},
+  volume        = {132},
+  pages         = {29--46},
+  year          = {2019},
+  publisher     = {Elsevier}
+}
+
+@article{simdicpp,
+  title         = {Vector Forward Mode Automatic Differentiation on SIMD/SIMT architectures },
+  author        = {Jan Hückelheim and Michel Schanen and Sri Hari Krishna Narayanan and Paul Hovland},
+  journal       = {International Conference on Parallel Processing},
+  volume        = {49},
+  year          = {2020}
+}
+@article{pips,
+author = {Petra, Cosmin G. and Schenk, Olaf and Lubin, Miles and Gäertner, Klaus},
+title = {An Augmented Incomplete Factorization Approach for Computing the Schur Complement in Stochastic Optimization},
+journal = {SIAM Journal on Scientific Computing},
+volume = {36},
+number = {2},
+pages = {C139-C162},
+year = {2014},
+doi = {10.1137/130908737},
+}
+
+@article{wachter2004implementation,
[…]
