+++ paper/paper.tex
@@ -1,0 +1,312 @@
+
+% JuliaCon proceedings template
+\documentclass{juliacon}
+\setcounter{page}{1}
+\newcommand{\refalg}[1]{Algorithm~\ref{#1}}
+\newcommand{\refsec}[1]{Section~\ref{#1}}
+\newcommand{\reffig}[1]{Figure~\ref{#1}}
+\newcommand{\refsubfig}[1]{Figure~\subref{#1}}
+\newcommand{\reftab}[1]{Table~\ref{#1}}
+\newcommand{\refeqn}[1]{(\ref{#1})}
+\newcommand{\reflst}[1]{Listing~(\ref{#1})}
+
+\begin{document}
+
+\input{header}
+
+\maketitle
+
+\begin{abstract}
+
+Solving optimal power flow is an important tool in the secure and cost
+effective operation of the transmission power grids. \lstinline{ExaPF.jl} aims to
+implement a reduced space method for solving the optimal power flow problem (OPF)
+fully on GPUs. Reduced space methods enforce the constraints, represented here by
+the power flow's (PF) system of nonlinear equations, separately at each
+iteration of the optimization in the reduced space. This paper describes the
+API of \lstinline{ExaPF.jl} for solving the power flow's nonlinear equations entirely on the GPU.
+This includes the computation of the derivatives using automatic
+differentiation, an iterative linear solver with a preconditioner, and a
+Newton-Raphson implementation. All of these steps allow us to run the main
+computational loop entirely on the GPU with no transfer from host to device.
+
+This implementation will serve as the basis for the future OPF implementation
+in the reduced space.
+
+\end{abstract}
+
+\section{Statement of Need}
+
+The current state-of-the-art for solving optimal power flow is the
+interior-point method (IPM) in optimization implemented by the solver Ipopt
+\cite{wachter2004implementation} and is the algorithm of reference
+in implementations like \verb MATPOWER \cite{matpower}. However, its reliance on
+unstructured sparse indefinite inertia revealing direct linear solvers makes
+this algorithm hard to port to GPUs. `ExaPF.jl` aims at applying a reduced
+gradient method to tackle this problem, which allows us to leverage iterative
+linear solvers for solving the linear systems arising in the PF.
+
+Our final goal is a reduced method optimization solver that provides a
+flexible API for models and formulations outside of the domain of OPF.
+
+\section{Components}
+
+To make our implementation portable to CPU and GPU architectures we leverage
+two abstractions: arrays and kernels. Both of these abstractions are
+supported through the packages \lstinline{CUDA.jl} \cite{besard2018juliagpu,besard2019prototyping,KernelAbstractions.jl}
+
+\subsection{AutoDiff}
+
+Given a set of equations \lstinline{F(x) = 0}, the Newton-Raphson algorithm for
+solving nonlinear equations (see below) requires the Jacobian \lstinline{J = jacobian(x)}
+of \lstinline{F}. At each iteration a new step \lstinline{dx} is computed by
+solving a linear system. In our case \lstinline{J} is sparse and indefinite, but
+invertible.
+
+\begin{lstlisting}[language = Julia]
+  go = true
+  while(go)
+    dx .= jacobian(x)\F(x)
+    x  .= x .- dx
+    go = norm(f(x)) < tol ? true : false
+  end
+\end{lstlisting}
+
+There are two modes of differentiation called *forward/tangent* or
+*reverse/adjoint*. The latter is known in machine learning as
[…]
