### 2021-01-29 #65 @crstnbr
@michel2323 You need to drop the `paper.md` file for PDF compilation to work. It clashes with the `paper.tex` file.
