+++ paper/paper.yml
@@ -19,4 +19,4 @@ affiliations:
  - name: Argonne National Laboratory
    index: 1
 date: 30 September 2020
-bibliography: paper.bib
+bibliography: ref.bib
