### 2021-01-29 #65 @michel2323
@crstnbr Thanks for the help. @vchuravy Here we are! Back then we followed the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html) when writing the paper. Not sure that is what you expect. Let us know what you think.
