### 2021-01-29 #65 @vchuravy
I noticed in the draft on page 2 that two references are not converted in Latex form.

@michel2323 Thanks! You can either decide to keep it this way or extend it to a 6-8 papers :)
