+++ paper/paper.tex
@@ -1,4 +1,3 @@
-
 % JuliaCon proceedings template
 \documentclass{juliacon}
 \setcounter{page}{1}
@@ -72,9 +71,9 @@ invertible.
   end
 \end{lstlisting}
 
-There are two modes of differentiation called *forward/tangent* or
-*reverse/adjoint*. The latter is known in machine learning as
-transposed Jacobian-vector product \lstinline{adj(x,y) = (J(x)'*y)}. We recommend
+There are two modes of differentiation called {\it forward/tangent} or
+{\it reverse/adjoint}. The latter is known in machine learning as
+transposed Jacobian-vector product \lstinline{adj(x,y) = J(x)'*y}. We recommend
 \cite{griewank2008evaluating} for a more in-depth introduction to automatic
 differentiation. The computational complexity of both models favors the
 adjoint mode if the number of outputs of \lstinline{F} is much smaller than the
@@ -100,8 +99,8 @@ independent columns in one Jacobian-vector evaluation (see
 algorithm implemented by \lstinline{SparseDiffTools.jl} \cite{sparsedifftools}.
 
 Given the sparsity pattern, the forward model is applied through the package
-\lstinline{ForwardDiff.jl} \cite{RevelsLubinPapamarkou2016}. Given the number of Jacobian
-colors $c$ we can build our dual type \lstinline{t1s} with \lstinline{c} directions:
+\lstinline{ForwardDiff.jl} \cite{RevelsLubinPapamarkou2016}. With the number of Jacobian
+colors $c$ we can build our dual type \lstinline{t1s} with \lstinline{N=c} directions:
 
 \begin{lstlisting}[language = Julia]
 t1s{N} = 
@@ -133,9 +132,9 @@ or accessed in kernels written with `KernelAbstractions.jl`, like for example th
 
 \begin{lstlisting}[language = Julia]
 @kernel function residual_kernel!(F, v_m, v_a,
-                    re_nzval, re_colptr, re_rowval,
-                    im_nzval, im_colptr, im_rowval,
-                    pinj, qinj, pv, pq, nbus)
+                re_nzval, re_colptr, re_rowval,
+                im_nzval, im_colptr, im_rowval,
+                pinj, qinj, pv, pq, nbus)
 
     npv = size(pv, 1)
     npq = size(pq, 1)
@@ -197,9 +196,9 @@ Our package supports the following linear solvers:
 \begin{itemize}
     \item CUSOLVER with `csrlsvqr` (GPU),
     \item `Krylov.jl` with `dqgmres` (CPU/GPU),
-    \item `IterativeSolvers` with `bicgstab` (CPU) [@sleijpen1993bicgstab],
+    \item `IterativeSolvers` with `bicgstab` (CPU) \cite{sleijpen1993bicgstab},
     \item UMFPACK through the default Julia `\` operator (CPU),
-    \item and a custom BiCGSTAB implementation [@bicgstabVorst] (CPU/GPU).
+    \item and a custom BiCGSTAB implementation \cite{bicgstabVorst} (CPU/GPU).
 \end{itemize}
 
 The last custom implementation was necessary as BiCGSTAB showed much better
@@ -234,7 +233,7 @@ Assuming that other vendors will provide such batched BLAS APIs, this code is po
 
 \section{Performance Example}
 
-To get an impression of the use case for this solver we consider a 30,000 bus system case from the ARPA-E GO competition.
+To illustrate the use case for this solver we consider a 30,000 bus system case from the ARPA-E GO competition.
 We show how the convergence of the BiCGSTAB algorithm is impacted by the number of blocks in the Jacobi preconditioner.
 By choosing appropriately the number of blocks, we observe that the iterative solver
 takes 0.55s to solve the linear system on the GPU. As a comparison,
