### 2021-02-04 #65 @michel2323
@vchuravy I fixed the references and did a last pass with a few minor beautifications. I added some keywords in `jlcode.sty`. I hope that's okay,

We'd go for short 3 pages and keep it this way.

Wheadon is suspiciously quiet.
