### 2021-02-04 #65 @crstnbr
> I added some keywords in `jlcode.sty`. I hope that's okay,

You shouldn't modify the template files. Only `paper.tex`, `ref.bib`, and `paper.yml`.

> @wheadon generate pdf from branch paper

It is "whedon" not "wheadon".
