+++ paper/paper.yml
@@ -9,15 +9,15 @@ keywords:
 authors:
   - name: Michel Schanen
     orcid: 0000-0002-4164-027X
-    affiliation: 1
+    # affiliation: 1
   - name: Adrian Maldonado
-    affiliation: 1
+    # affiliation: 1
   - name: François Pacaud
-    affiliation: 1
+    # affiliation: 1
   - name: Mihai Anitescu
-    affiliation: 1
+    # affiliation: 1
 affiliations:
  - name: Argonne National Laboratory
-   index: 1
+  #  index: 1
 date: 30 September 2020
 bibliography: ref.bib
