### 2021-03-25 #72 @matbesancon
hi @michel2323, it looks like you have some DOIs missing on the paper, could you see if the suggested missing DOIs apply to the references you cite
