### 2021-03-29 #72 @lcw
In addition to reading the submitted paper, I checked out the code from
https://github.com/exanauts/ExaPF.jl and tried to run it.

The submitted paper gives a short description of the need for optimal power flow
solvers on the GPU, a detailed description of the technical components (autodiff
and linear solvers), and finishes with a couple performance results.  In all
this is an ambitious project and I applaud the authors efforts.  It is exciting to
see more and more code in the julia ecosystem work for both the CPU and GPU.

Please find below a list of comments and concerns I have with the current
presentation of the work.

  - I submitted various issues and a pr during this review and the authors were
    respectful, helpful, and prompt in their responses.
  - Initially the test suite (for v0.4.0) didn't work for me,
    https://github.com/exanauts/ExaPF.jl/pull/136, as I was using too new of a
    julia version.  Apparently, this was already fixed on the development
    branch, https://github.com/exanauts/ExaPF.jl/pull/136#issuecomment-808632264.
    It would be helpful if there were community guidelines describing the
    development process for the code, https://github.com/exanauts/ExaPF.jl/issues/139.
  - I experienced non-deterministic failures in the test suite
    https://github.com/exanauts/ExaPF.jl/issues/137 suggesting a possible race
    condition in their GPU code.
  - The paper claims to describe the API of ExaPF.jl but there is no discussion
    of the functions a user would typically call.  So, I looked to the quick
    start document to see how to exercise the code but unfortunately it has not
    been kept up to date in v0.4.0 or the development branch,
    https://github.com/exanauts/ExaPF.jl/issues/138.
  - I was unable to reproduce the performance results described in the paper.
    There is a `benchmark` directory but a lack of documentation on what exactly
    was run inhibited my pursuit.
  - I am glad to see that the authors included performance results.  However, I was
    surprised to see that the CPU code was faster.  This begs the question, why
    go through all this effort to make the code run on the GPU if it is not
    faster?

# Conflict of interest

- [X] As the reviewer I confirm that I have read the JuliaCon conflict of interest policy and that there are no conflicts of interest for me to review this work.

# Code of Conduct

- [X] I confirm that I read and will adhere to the JuliaCon code of conduct.

# General checks

- [X] Repository: Is the source code for this software available at the repository url?
- [X] License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
- [X] Authorship: Has the submitting author (@michel2323) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

- [X] Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

# Paper format

- [X] Authors: Does the paper.tex file include a list of authors with their affiliations?
- [ ] References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?
- [ ] Page limit: Is the page limit for full papers respected by the submitted document?

# Content
- [ ] Context: is the scientific context motivating the work correctly presented?
- [ ] Methodology: is the approach taken in the work justified, presented with enough details and reference to reproduce it?
- [ ] Results: are the results presented and compared to approaches with similar goals?
