### 2021-04-01 #72 @frapac
Dear all,

First, we would like to thank the reviewer for his thorough and meaningful review. To our concern, all points raised by the reviewer are relevant.

We are currently addressing the most urgent tasks, as
- improving user friendliness (mostly with the documentation).
- adding encapsulated scripts (with proper Project.toml and Manifest.toml) to reproduce the results presented in the paper.
- improving the README.md file, as an entry point to ExaPF.
- however, the race condition on the GPU would be harder to debunk. We are facing this issue for a few months, and haven't been able to isolate it properly (yet).

Once these points are addressed, we are planning to proceed to a new release for ExaPF (v0.5.0).

Remains the last (but not least) concern raised by the reviewer, pointing to the performance of the GPU code versus the CPU code. We acknowledge that this claim is highlt legitimate: why bothering porting code on the GPU if overall the CPU code is faster? We give two answers to that concern, depending whether we look at things from the top or from the bottom up.

- The code is designed for and funded by the requirement of running on GPU centric architectures where CPU nodes do only provide a negligible amount of computational performance (Summit at OLFC, 46 TFLOPS by GPUs vs 1 TFLOP CPU per node. Furthermore, in the electrical engineering community or the optimization community, most codes are heavily optimized towards the CPU. Developing code on the GPU requires careful thinking, as we must revisit most of the algorithms developed during the last decades. However, we have hope that ExaPF will soon be able to beat the CPU implementation (see points below).
- Looking more closely at the performance, we have observed that on the GPU the bottleneck of the powerflow solver is the (sparse) linear solver. The performance offered by default by CUSOLVER lags behind its CPU equivalents (like UMFPACK). We are currently addressing this issue by replacing CUSOLVER by [CUSOLVER_RF](https://docs.nvidia.com/cuda/cusolver/index.html#cuSolverRF-reference) (currently not interfaced in CUDA.jl), which allows to store the LU factorization used throughout the powerflow algorithm. With this latest development, coupled with efficient batch solves, we think we have all the ingredients in hand to beat the CPU implementation.

We hope that these two points allowed us to clarify some of the concerns of the reviewer.

Eventually, this raises the question of the current state of the article, whose results are lagging behind the most recent developments. The proceedings paper is based on the release as of JuliaCon 2020 and ran on Julia 1.4, and does not cover the latest development made in ExaPF.  That’s the reason why we would like to propose two paths onward:
1. either we address the concerns of the reviewer and leave the results unchanged, representing the status of submission of JuliaCon 2020 with Julia 1.4.
2. or we withdraw the paper as is, and present a revamped version for JuliaCon 2021.

Michel Schanen & François Pacaud
