### 2021-04-01 #72 @matbesancon
Thanks François for the detailed answer. I would be okay with both paths onwards and will leave it to the authors to choose. One thing to note: the conference is continuously publishing the proceedings, so there is no need to choose specifically "publishing in JuliaCon 2020 or 2021". Given that, I don't think it is worth it to withdraw to resubmit, it would be the same thing as resubmitting. So specifically, I am okay with keeping the results are you ran them on the original submission if you prefer.

The second point of answer you mentioned above "Looking more closely at the performance..." offers great insight that is worth having in the paper IMO
