+++ paper/ref.bib
@@ -24,7 +24,9 @@ doi = {10.1137/0913035},
   year={2017},
   volume={8},
   number={3},
-  pages={1378-1386},}
+  pages={1378-1386},
+  doi={10.1109/tsg.2016.2610863}
+  }
 
 @article{RevelsLubinPapamarkou2016,
     title = {Forward-Mode Automatic Differentiation in {J}ulia},
@@ -53,7 +55,8 @@ doi = {10.1137/0913035},
   volume        = {132},
   pages         = {29--46},
   year          = {2019},
-  publisher     = {Elsevier}
+  publisher     = {Elsevier},
+  doi           = {10.1016/j.advengsoft.2019.02.002}
 }
 
 @article{simdicpp,
@@ -61,7 +64,8 @@ doi = {10.1137/0913035},
   author        = {Jan Hückelheim and Michel Schanen and Sri Hari Krishna Narayanan and Paul Hovland},
   journal       = {International Conference on Parallel Processing},
   volume        = {49},
-  year          = {2020}
+  year          = {2020},
+  doi           = {10.1145/3404397.3404470}
 }
 @article{pips,
 author = {Petra, Cosmin G. and Schenk, Olaf and Lubin, Miles and Gäertner, Klaus},
@@ -82,7 +86,8 @@ doi = {10.1137/130908737},
   number={1},
   pages={25--57},
   year={2006},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s10107-004-0559-y}
 }
 
 @ARTICLE{matpower,
@@ -93,6 +98,7 @@ doi = {10.1137/130908737},
   volume={22},
   number={3},
   pages={1185-1193},
+  doi={10.1109/tpwrs.2007.901301}
 }
 @book{griewank2008evaluating,
   title={Evaluating derivatives: principles and techniques of algorithmic differentiation},
