### 2021-04-22 #72 @matbesancon
there seem to be an issue with the references
