### 2021-04-22 #72 @matbesancon
@michel2323 you should use the same bibtex structure left as-is: comment out the bibliographystyle and bibliography insertion and uncomment the ones from below
