+++ paper/paper.tex
@@ -267,8 +267,8 @@ This shows the number of BiCGSTAB iterations and the time needed to achieve conv
 
 This research was supported by the Exascale Computing Project (17-SC-20-SC), a joint project of the U.S. Department of Energy’s Office of Science and National Nuclear Security Administration, responsible for delivering a capable exascale ecosystem, including software, applications, and hardware technology, to support the nation’s exascale computing imperative.
 
-\bibliographystyle{juliacon}
-\bibliography{paper}
+% \bibliographystyle{juliacon}
+% \bibliography{paper}
 
 % \section{Additional features}
 % \label{sec:additional_faci}
@@ -313,7 +313,7 @@ This research was supported by the Exascale Computing Project (17-SC-20-SC), a j
 % parties, the ref.bib file should be sent with it.
 % \cite{bezanson2017julia}
 
-% \input{bib.tex}
+\input{bib.tex}
 
 \end{document}
 
