### 2021-06-03 #72 @matbesancon
Thanks the reference issue seems fixed now.
@lcw can you confirm the new changes correspond to your previous comments and whether the manuscript is now good on your side?
