### 2021-06-08 #72 @lcw
The authors have addressed many of my issues and I understand that some of the ones I brought up are out of scope.  The readme is greatly improved, however, I am unable to run the examples https://github.com/exanauts/ExaPF.jl#how-to-solve-the-power-flow-of-a-given-matpower-instance and https://github.com/exanauts/ExaPF.jl#how-to-solve-the-optimal-power-flow-in-the-reduced-space without modification.

That said, I am okay if the authors want to push forward and publish.
