### 2021-07-01 #72 @matbesancon
@michel2323 @frapac do you want to address the points above on the examples? Otherwise let us know if you did the changes you wanted
