### 2021-07-16 #72 @frapac
Dear Mathieu,
First, I am sorry for the late reply. We are planning to address the two last points raised  by Lucas by the JuliaCon, to get everything working by then. Then, we will update the JuliaCon paper with our latest results, as discussed earlier in this discussion. Would that be ok for you?
